+++ libs/stack/stack-ui/src/providers/Alerts/interface.ts
@@ -1,0 +1,11 @@
+import type { ReactNode } from 'react'
+import type { SwiperClass } from 'swiper/react'
+
+export interface TAlertsControllerProviderProps {
+  children: ReactNode
+  controller: SwiperClass | undefined
+}
+
+export type TAlertsControllerContext = {
+  controller: SwiperClass | undefined
+}